requirements/dev.txt
-r base.txt
pytest>=8.3
pytest-django>=4.8
pytest-cov>=5.0
factory-boy>=3.3
model-bakery>=1.19
black>=24.10
isort>=5.13
flake8>=7.1
bandit>=1.7
django-debug-toolbar>=4.4
